Fix the bugs for edit code with AI
