+++ src/completion.ts
@@ -1,5 +1,5 @@
 import {Application} from "./application";
-import {LlamaResponse} from "./llama-server";
+import { LlamaResponse } from "./types";
 import vscode from "vscode";
 import {Utils} from "./utils";
