+++ src/types.ts
@@ -1,3 +1,28 @@
+export interface ChatMessage {
+  role: string; // or just 'string' if you need more roles
+  content: string;
+  tool_call_id?: string
+}
+
+export interface LlamaResponse {
+    content?: string;
+    generation_settings?: any;
+    tokens_cached?: number;
+    truncated?: boolean;
+    timings?: {
+        prompt_n?: number;
+        prompt_ms?: number;
+        prompt_per_second?: number;
+        predicted_n?: number;
+        predicted_ms?: number;
+        predicted_per_second?: number;
+    };
+}
+
+export interface LlamaChatResponse {
+    choices: [{message:{content?: string}}];
+}
+
 export interface ChunkEntry {
     uri: string;
     content: string;
